requirements.txt
# Core ML and NLP
torch==2.1.0
transformers==4.36.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
spacy==3.7.2
textblob==0.17.1
nltk==3.8.1

# Flask and Web
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Limiter==3.3.1
Flask-Talisman==1.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
Flask-Redis==0.4.0
Werkzeug==2.3.7
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.38.3
psycopg2-binary==2.9.7
redis==5.0.1

# Document processing
PyPDF2==3.0.1
python-docx==1.0.1
openpyxl==3.1.2

# Monitoring and Metrics
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.2
python-json-logger==2.0.7
psutil==5.9.6

# Security
bcrypt==4.0.1
argon2-cffi==21.3.0
PyJWT==2.8.0

# Utilities
tqdm==4.66.2
huggingface-hub>=0.16.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1
colorama==0.4.6
pyyaml==6.0.1
ujson==5.8.0

# Development and Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0
coverage==7.3.2